Rename files to kebabcase
