+++ packages/substate/src/index.js
@@ -2,4 +2,4 @@ export { default as createStore } from 'zustand';
 export { default as shallowCompare } from 'zustand/shallow';
 export { default as createBaseStore } from 'zustand/vanilla';
 
-export * from './useSubState';
+export * from './use-sub-state';
